Add files via upload
Move - Java Program to Move All 'x' Characters to the End Using Recursion
Overview
This Java program moves all occurrences of the character 'x' in a given string to the end while maintaining the order of other characters. It uses recursion to efficiently solve the problem.

How It Works
Recursive Function (moveAllX)

Base case: If idx == str.length(), append count number of 'x' characters to newString and print it.
If the current character is 'x', increment the count and continue recursion.
Otherwise, add the character to newString and recurse further.
Main Function

Calls moveAllX(str, 0, 0, "") with an example string ("axbcxxd").